+++ frontend/app/(app)/trips/[id]/new-activity-tab/components/activity-context-menu.tsx
@@ -1,48 +1,17 @@
 import { Box, Text } from "@/design-system";
+import GlassContainer from "@/design-system/components/glass-container";
 import { ColorPalette } from "@/design-system/tokens/color";
-import { CornerRadius } from "@/design-system/tokens/corner-radius";
-import { Elevation } from "@/design-system/tokens/elevation";
-import { Layout } from "@/design-system/tokens/layout";
 import { BorderWidth } from "@/design-system/tokens/border";
-import { LinearGradient } from "expo-linear-gradient";
+import { Layout } from "@/design-system/tokens/layout";
 import { CheckCircle, ExternalLink } from "lucide-react-native";
-import { Modal, Pressable, StyleSheet, View } from "react-native";
-
-// ─── Types ───────────────────────────────────────────────────────────────────
-
-type MenuPosition = { top: number; right: number };
-
-type ActivityContextMenuProps = {
-  visible: boolean;
-  goToLabel: string;
-  anchorPosition: MenuPosition | null;
-  onClose: () => void;
-  onMarkRead?: () => void;
-  onGoTo?: () => void;
-};
-
-// ─── Constants ───────────────────────────────────────────────────────────────
-
-const ICON_SIZE = 18;
-const MENU_WIDTH = 214;
-
-// Glass effect colors
-const GLASS_BG = "rgba(206, 224, 238, 0.42)";
-const GLASS_BORDER = "rgba(255, 255, 255, 0.9)";
-const GLASS_GRADIENT = [
-  "rgba(246, 249, 252, 0.92)",
-  "rgba(194, 220, 238, 0.62)",
-  "rgba(186, 212, 230, 0.5)",
-  "rgba(238, 236, 213, 0.3)",
-  "rgba(170, 194, 214, 0.48)",
-] as const;
-const GLASS_GRADIENT_STOPS = [0, 0.28, 0.62, 0.76, 1] as const;
-const SHINE_GRADIENT = [
-  "rgba(255, 255, 255, 0.86)",
-  "rgba(255, 255, 255, 0.04)",
-] as const;
-const DIVIDER_COLOR = "rgba(41, 54, 68, 0.2)";
-const ROW_PRESSED_BG = "rgba(255, 255, 255, 0.35)";
+import { Modal, Pressable, StyleSheet } from "react-native";
+import {
+  CONTEXT_MENU_ICON_SIZE,
+  CONTEXT_MENU_WIDTH,
+  DIVIDER_COLOR,
+  ROW_PRESSED_BG,
+} from "../constants";
+import type { ActivityContextMenuProps } from "../types";
 
 // ─── Component ───────────────────────────────────────────────────────────────
 
@@ -71,27 +40,11 @@ export default function ActivityContextMenu({
             {
               top: anchorPosition.top + Layout.spacing.xxs,
               right: anchorPosition.right,
-              width: MENU_WIDTH,
+              width: CONTEXT_MENU_WIDTH,
             },
           ]}
         >
-          <View style={styles.menuContainer}>
-            <LinearGradient
-              colors={[...GLASS_GRADIENT]}
-              locations={[...GLASS_GRADIENT_STOPS]}
-              start={{ x: 0, y: 0 }}
-              end={{ x: 0, y: 1 }}
-              style={StyleSheet.absoluteFillObject}
-            />
-            <LinearGradient
-              colors={[...SHINE_GRADIENT]}
-              locations={[0, 1]}
-              start={{ x: 0.5, y: 0 }}
-              end={{ x: 0.5, y: 1 }}
-              style={styles.topShine}
-            />
-            <View style={styles.glossLayer} />
-
+          <GlassContainer>
             {onMarkRead && (
               <>
                 <Pressable
@@ -104,7 +57,7 @@ export default function ActivityContextMenu({
                     pressed && styles.rowPressed,
                   ]}
                 >
-                  <CheckCircle size={ICON_SIZE} color={ColorPalette.gray900} />
+                  <CheckCircle size={CONTEXT_MENU_ICON_SIZE} color={ColorPalette.gray900} />
                   <Text variant="bodySmMedium" color="gray900">
                     Mark as read
                   </Text>
@@ -124,13 +77,13 @@ export default function ActivityContextMenu({
                   pressed && styles.rowPressed,
                 ]}
               >
-                <ExternalLink size={ICON_SIZE} color={ColorPalette.gray900} />
+                <ExternalLink size={CONTEXT_MENU_ICON_SIZE} color={ColorPalette.gray900} />
                 <Text variant="bodySmMedium" color="gray900">
                   {goToLabel}
                 </Text>
               </Pressable>
             )}
-          </View>
+          </GlassContainer>
         </Pressable>
       </Pressable>
     </Modal>
@@ -146,22 +99,6 @@ const styles = StyleSheet.create({
   },
   menuAnchor: {
     position: "absolute",
-  },
-  menuContainer: {
-    overflow: "hidden",
-    borderRadius: CornerRadius.xxl,
-    borderWidth: BorderWidth.thin,
-    borderColor: GLASS_BORDER,
-    backgroundColor: GLASS_BG,
-    ...Elevation.xl,
-  },
-  topShine: {
-    ...StyleSheet.absoluteFillObject,
-    height: "42%",
-  },
-  glossLayer: {
-    ...StyleSheet.absoluteFillObject,
-    backgroundColor: "rgba(255, 255, 255, 0.05)",
   },
   row: {
     minHeight: 54,
